#32639: Make IPTR unsigned long

+++ gcc/testsuite/g++.target/riscv/tt/sfpi/conv-qsr.C
@@ -2,7 +2,7 @@
 // { dg-final { check-function-bodies "**" "" } }
 
 namespace ckernel{
-    unsigned *instrn_buffer;
+    unsigned long *instrn_buffer;
 }
 #include <sfpi.h>
