+++ gcc/testsuite/g++.target/riscv/tt/sfpi/nodef-40182-bh.C
@@ -1,7 +1,7 @@
 // { dg-options "-mcpu=tt-bh-tensix -O2 -I [SFPI]/include -fno-exceptions -fno-rtti -Wall -Wextra -Wunused-parameter" }
 
 namespace ckernel{
-    extern unsigned instrn_buffer[];
+    extern unsigned long instrn_buffer[];
 }
 #include <lltt.h>
 #include <sfpi.h>
